feat: functions for common functional operations
Implemented functions for the purpose of generating new molecules via
functional group substitutions. The goal is to build a general interface
that allows the user to substitute substructures with others that are
commonly found in astrochemistry.

Planning ahead a little as well, it would be nice to build in a way of
recursively generating these structures, particularly for the generation
of PAHs.