feat: implemented a tree structure for molecule substitution
Implemented a class that abstracts out simple molecule graph in a tree
like manner using MoleculeGenerator and MoleculeNode classes.

For the former, the main form of interaction is a `grow_tree` method
which will randomly grow the tree for a specified number of iterations.
Each iteration will randomly select a node, a substructure to match and
to replace, and use the `replace_substructure` function to exhaustively
generate all new replacements.